fix : usage of correct jdbc sql mapping type for datatype "Long" with null - value

+++ jodd-db/src/main/java/jodd/db/DbQuery.java
@@ -349,7 +349,7 @@ public class DbQuery<Q extends DbQuery> extends DbQueryBase<Q> {
 
 	public Q setLong(int index, Number value) {
 		if (value == null) {
-			setNull(index, Types.INTEGER);
+			setNull(index, Types.BIGINT);
 		}
 		else {
 			setLong(index, value.longValue());
@@ -359,7 +359,7 @@ public class DbQuery<Q extends DbQuery> extends DbQueryBase<Q> {
 
 	public Q setLong(String param, Number value) {
 		if (value == null) {
-			setNull(param, Types.INTEGER);
+			setNull(param, Types.BIGINT);
 		}
 		else {
 			setLong(param, value.longValue());
